Make an empty StringSet nil.
